requirements.txt
absl-py==2.1.0
ai21==2.15.1
ai21-tokenizer==0.12.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
altair==5.4.1
aniso8601==9.0.1
annotated-types==0.7.0
anthropic==0.36.1
anyio==4.6.2.post1
astunparse==1.6.3
attrs==23.2.0
blinker==1.8.2
blis==0.7.11
boto3==1.35.41
botocore==1.35.41
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==2.2.1
cohere==5.11.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
databricks-sdk==0.34.0
Deprecated==1.2.14
dill==0.3.9
distro==1.9.0
docker==7.1.0
eval_type_backport==0.2.0
fastapi==0.115.2
fastavro==1.9.7
ffmpy==0.4.0
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.65.0
gradio==4.26.0
gradio_client==0.15.1
graphene==3.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
groq==0.11.0
grpcio==1.67.0
grpcio-status==1.62.3
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.2
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
itsdangerous==2.2.0
jellyfish==1.1.0
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.6.0
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
libclang==18.1.1
Mako==1.3.5
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mistralai==1.1.0
ml-dtypes==0.4.1
mlflow==2.17.0
mlflow-skinny==2.17.0
mock==4.0.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
murmurhash==1.0.10
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.9.3
networkx==3.4.1
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.51.2
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opt_einsum==3.4.0
optree==0.13.0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
parameterized==0.9.0
pathos==0.3.3
pillow==10.4.0
platformdirs==4.3.6
pox==0.3.5
ppft==1.7.6.9
preshed==3.0.9
propcache==0.2.0
proto-plus==1.24.0
protobuf==4.25.5
psutil==6.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
pywin32==308
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
ruff==0.6.9
s3transfer==0.10.3
safetensors==0.4.5
sagemaker==2.232.2
sagemaker-core==1.0.10
sagemaker-mlflow==0.1.0
schema==0.7.7
scikit-learn==1.5.2
scipy==1.10.0
segtok==1.5.11
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
shuttleai==3.9.6
six==1.16.0
smart-open==6.4.0
smdebug-rulesconfig==1.0.1
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
sqlparse==0.5.1
srsly==2.4.8
starlette==0.40.0
swarmauri==0.4.1
sympy==1.13.3
tabulate==0.9.0
tblib==3.0.0
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
textblob==0.18.0.post0
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.20.1
tomlkit==0.12.0
torch==2.4.1
tqdm==4.66.5
transformers==4.45.2
typer==0.9.4
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
waitress==3.0.0
wasabi==1.1.3
weasel==0.3.4
websockets==11.0.3
Werkzeug==3.0.4
wrapt==1.16.0
yake==0.4.8
yarl==1.15.3
zipp==3.20.2
